feat: 修改Numpy数据类型设置方式

+++ code/chapter-8/04_tracking/deep_sort/deep_sort/deep_sort.py
@@ -49,7 +49,7 @@ class DeepSort(object):
             box = track.to_tlwh()
             x1, y1, x2, y2 = self._tlwh_to_xyxy(box)
             track_id = track.track_id
-            outputs.append(np.array([x1, y1, x2, y2, track_id], dtype=np.int))
+            outputs.append(np.array([x1, y1, x2, y2, track_id], dtype=np.int32))
         if len(outputs) > 0:
             outputs = np.stack(outputs, axis=0)
         return outputs
